rename mock fetch response
