+++ src/api/utils/handleResponse.spec.ts
@@ -1,19 +1,19 @@
 import APIError from 'utils/apiError';
 import handleResponse from './handleResponse';
-import mockZetkinResponse from 'test-utils/mocks/mockZetkinResult';
+import mockFetchResponse from 'test-utils/mocks/mockFetchResponse';
 
 describe('handleResponse()', () => {
 
     it('Returns the value at the property "data" of the response object', async () => {
         const resBody = { data: { id: 1, title: 'Coffee Table book about coffee tables' } };
-        const mock200Res = mockZetkinResponse(resBody);
+        const mock200Res = mockFetchResponse(resBody);
         const data = await handleResponse(mock200Res, 'GET');
         expect(data).toEqual(resBody.data);
     });
 
     describe('Throws errors', () => {
         it('when there is a non-OK status code', async () => {
-            const mock404Res = mockZetkinResponse({}, { ok: false , status: 404 });
+            const mock404Res = mockFetchResponse({}, { ok: false , status: 404 });
             const handle = () => handleResponse(mock404Res, 'GET');
             expect.assertions(1);
             return handle().catch(e =>
@@ -22,7 +22,7 @@ describe('handleResponse()', () => {
         });
 
         it('when the response object does not contain the property "data"', async () => {
-            const mock404Res = mockZetkinResponse({ notBody: 'this is not the property body' });
+            const mock404Res = mockFetchResponse({ notBody: 'this is not the property body' });
             const handle = () => handleResponse(mock404Res, 'GET');
             expect.assertions(1);
             return handle().catch(e =>
